測定方法の変更

+++ cicada/include/common.hpp
@@ -16,7 +16,7 @@
 #ifdef GLOBAL_VALUE_DEFINE
   #define GLOBAL
 
-GLOBAL std::atomic<unsigned int> Running(0);
+GLOBAL std::atomic<size_t> Running(0);
 GLOBAL std::atomic<bool> Finish(false);
 alignas(CACHE_LINE_SIZE) GLOBAL std::atomic<uint64_t> MinRts(0);
 alignas(CACHE_LINE_SIZE) GLOBAL std::atomic<uint64_t> MinWts(0);
@@ -25,7 +25,7 @@ alignas(CACHE_LINE_SIZE) GLOBAL std::atomic<unsigned int> FirstAllocateTimestamp
 #else
   #define GLOBAL extern
 
-GLOBAL std::atomic<unsigned int> Running;
+GLOBAL std::atomic<size_t> Running;
 GLOBAL std::atomic<bool> Finish;
 alignas(CACHE_LINE_SIZE) GLOBAL std::atomic<uint64_t> MinRts;
 alignas(CACHE_LINE_SIZE) GLOBAL std::atomic<uint64_t> MinWts;
@@ -33,21 +33,21 @@ alignas(CACHE_LINE_SIZE) GLOBAL std::atomic<unsigned int> FirstAllocateTimestamp
 
 #endif
 
-GLOBAL unsigned int TUPLE_NUM;
-GLOBAL unsigned int MAX_OPE;
-GLOBAL unsigned int THREAD_NUM;
-GLOBAL unsigned int RRATIO;
+GLOBAL size_t TUPLE_NUM;
+GLOBAL size_t MAX_OPE;
+GLOBAL size_t THREAD_NUM;
+GLOBAL size_t RRATIO;
 GLOBAL bool RMW;
 GLOBAL double ZIPF_SKEW;
 GLOBAL bool YCSB;
 GLOBAL bool P_WAL;
 GLOBAL bool S_WAL;
-GLOBAL unsigned int GROUP_COMMIT;
-GLOBAL uint64_t CLOCKS_PER_US; // US = micro(µ) seconds 
-GLOBAL double IO_TIME_NS; // nano second
-GLOBAL uint64_t GROUP_COMMIT_TIMEOUT_US; // micro seconds
-GLOBAL uint64_t GC_INTER_US; // garbage collection interval
-GLOBAL uint64_t EXTIME;
+GLOBAL size_t GROUP_COMMIT;
+GLOBAL size_t CLOCKS_PER_US; // US = micro(µ) seconds 
+GLOBAL size_t IO_TIME_NS; // nano second
+GLOBAL size_t GROUP_COMMIT_TIMEOUT_US; // micro seconds
+GLOBAL size_t GC_INTER_US; // garbage collection interval
+GLOBAL size_t EXTIME;
 
 GLOBAL uint64_t_64byte *ThreadWtsArray;
 GLOBAL uint64_t_64byte *ThreadRtsArray;
