+++ silo/include/common.hpp
@@ -14,27 +14,27 @@
 #ifdef GLOBAL_VALUE_DEFINE
   #define GLOBAL
 
-GLOBAL std::atomic<unsigned int> Running(0);
+GLOBAL std::atomic<size_t> Running(0);
 alignas(CACHE_LINE_SIZE) GLOBAL uint64_t_64byte GlobalEpoch(1);
 
 #else
   #define GLOBAL extern
 
-GLOBAL std::atomic<unsigned int> Running;
+GLOBAL std::atomic<size_t> Running;
 alignas(CACHE_LINE_SIZE) GLOBAL uint64_t_64byte GlobalEpoch;
 
 #endif
 
-GLOBAL unsigned int TUPLE_NUM;
-GLOBAL unsigned int MAX_OPE;
-GLOBAL unsigned int THREAD_NUM;
-GLOBAL unsigned int RRATIO;
+GLOBAL size_t TUPLE_NUM;
+GLOBAL size_t MAX_OPE;
+GLOBAL size_t THREAD_NUM;
+GLOBAL size_t RRATIO;
 GLOBAL bool RMW;
 GLOBAL double ZIPF_SKEW;
 GLOBAL bool YCSB;
-GLOBAL uint64_t CLOCKS_PER_US;
-GLOBAL uint64_t EPOCH_TIME;
-GLOBAL unsigned int EXTIME;
+GLOBAL size_t CLOCKS_PER_US;
+GLOBAL size_t EPOCH_TIME;
+GLOBAL size_t EXTIME;
 
 alignas(CACHE_LINE_SIZE) GLOBAL uint64_t_64byte *ThLocalEpoch;
 alignas(CACHE_LINE_SIZE) GLOBAL uint64_t_64byte *CTIDW;
